Move Security Analysis Provider implementation into specific module
