+++ security-analysis/pom.xml
@@ -25,6 +25,7 @@
 
     <modules>
         <module>security-analysis-api</module>
+        <module>security-analysis-default</module>
     </modules>
 
 </project>
