--- security-analysis/security-analysis-api/src/test/resources/com/powsybl/config/test/config.yml
@@ -1,2 +1,0 @@
-security-analysis:
-  default-impl-name: "SecurityAnalysisImpl"
