+++ security-analysis/security-analysis-default/src/main/java/com/powsybl/security/impl/DefaultSecurityAnalysisProvider.java
@@ -4,12 +4,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.security;
+package com.powsybl.security.impl;
 
 import com.google.auto.service.AutoService;
 import com.powsybl.computation.ComputationManager;
 import com.powsybl.contingency.ContingenciesProvider;
 import com.powsybl.iidm.network.Network;
+import com.powsybl.security.*;
 import com.powsybl.security.interceptors.SecurityAnalysisInterceptor;
 
 import java.util.List;
@@ -19,9 +20,9 @@ import java.util.concurrent.CompletableFuture;
  * @author Thomas Adam <tadam at silicom.fr>
  */
 @AutoService(SecurityAnalysisProvider.class)
-public class SecurityAnalysisImplProvider implements SecurityAnalysisProvider {
+public class DefaultSecurityAnalysisProvider implements SecurityAnalysisProvider {
 
-    private static final String PROVIDER_NAME = "SecurityAnalysisImpl";
+    private static final String PROVIDER_NAME = "DefaultSecurityAnalysis";
     private static final String PROVIDER_VERSION = "1.0";
 
     @Override
@@ -33,7 +34,7 @@ public class SecurityAnalysisImplProvider implements SecurityAnalysisProvider {
                                                          SecurityAnalysisParameters parameters,
                                                          ContingenciesProvider contingenciesProvider,
                                                          List<SecurityAnalysisInterceptor> interceptors) {
-        SecurityAnalysisImpl securityAnalysis = new SecurityAnalysisImpl(network, detector, filter, computationManager);
+        DefaultSecurityAnalysis securityAnalysis = new DefaultSecurityAnalysis(network, detector, filter, computationManager);
         interceptors.forEach(securityAnalysis::addInterceptor);
         return securityAnalysis.run(workingVariantId, parameters, contingenciesProvider);
     }
