+++ security-analysis/security-analysis-default/src/test/java/com/powsybl/security/impl/interceptors/SecurityAnalysisInterceptorMock.java
@@ -4,9 +4,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.security.interceptors;
+package com.powsybl.security.impl.interceptors;
 
 import com.powsybl.security.*;
+import com.powsybl.security.interceptors.DefaultSecurityAnalysisInterceptor;
+import com.powsybl.security.interceptors.SecurityAnalysisResultContext;
 
 import static org.junit.Assert.*;
