+++ security-analysis/security-analysis-default/src/test/java/com/powsybl/security/impl/interceptors/SecurityAnalysisInterceptorMockExtension.java
@@ -4,9 +4,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.security.interceptors;
+package com.powsybl.security.impl.interceptors;
 
 import com.google.auto.service.AutoService;
+import com.powsybl.security.interceptors.SecurityAnalysisInterceptor;
+import com.powsybl.security.interceptors.SecurityAnalysisInterceptorExtension;
 
 /**
  * @author Mathieu Bague <mathieu.bague at rte-france.com>
