+++ security-analysis/security-analysis-default/src/test/java/com/powsybl/security/impl/mocks/LoadFlowProviderMock.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.security;
+package com.powsybl.security.impl.mocks;
 
 import com.google.auto.service.AutoService;
 import com.powsybl.computation.ComputationManager;
