+++ security-analysis/security-analysis-default/src/test/resources/com/powsybl/config/test/config.yml
@@ -1,0 +1,2 @@
+security-analysis:
+  default-impl-name: "DefaultSecurityAnalysis"
